VPP_Algorithm-main/requirements-lock.txt
# VPP算法服务生产环境依赖锁定文件 v2.0
# ============================================
# 
# 此文件锁定了所有依赖的具体版本，确保生产环境的一致性和安全性
# 使用方法: pip install -r requirements-lock.txt
# 
# 版本: 2.0.0
# 更新: 2024-12-29

# =============================================================================
# 🎯 核心算法依赖
# =============================================================================

# 数值计算和科学计算
numpy==1.24.4                    # 数值计算基础库
pandas==2.0.3                    # 数据处理和分析
scipy==1.11.4                    # 科学计算库
scikit-learn==1.3.2              # 机器学习算法

# 优化求解器
cvxpy==1.4.1                     # 凸优化求解器
mosek==10.1.24                   # 商业优化求解器（可选）

# 可视化（开发和调试用）
matplotlib==3.7.4                # 图表绘制
seaborn==0.13.0                  # 统计可视化

# =============================================================================
# 🌐 Web框架和API
# =============================================================================

# FastAPI生态
fastapi==0.104.1                 # 现代Web框架
uvicorn[standard]==0.24.0         # ASGI服务器
pydantic==2.5.0                  # 数据验证和序列化
pydantic[email]==2.5.0           # 邮箱验证支持

# HTTP客户端
httpx==0.25.2                    # 异步HTTP客户端
requests==2.31.0                 # 同步HTTP客户端

# 速率限制
slowapi==0.1.9                   # FastAPI速率限制

# =============================================================================
# 🗄️ 数据库和缓存
# =============================================================================

# PostgreSQL
psycopg2-binary==2.9.9           # PostgreSQL适配器
SQLAlchemy==2.0.23               # ORM框架
alembic==1.13.1                  # 数据库迁移工具

# Redis缓存
redis==5.0.1                     # Redis客户端
hiredis==2.2.3                   # 高性能Redis解析器

# =============================================================================
# 🔐 安全和认证
# =============================================================================

# JWT和加密
PyJWT==2.8.0                     # JWT令牌处理
cryptography==41.0.7             # 加密库
passlib[bcrypt]==1.7.4           # 密码哈希
python-multipart==0.0.6          # 表单数据解析

# =============================================================================
# 📊 监控和日志
# =============================================================================

# 日志和监控
structlog==23.2.0                # 结构化日志
prometheus-client==0.19.0        # Prometheus监控指标
opentelemetry-api==1.21.0        # 分布式追踪

# =============================================================================
# 🛠️ 开发和测试工具
# =============================================================================

# 测试框架
pytest==7.4.3                    # 测试框架
pytest-asyncio==0.21.1           # 异步测试支持
pytest-cov==4.1.0                # 测试覆盖率
pytest-mock==3.12.0              # Mock测试工具
pytest-xdist==3.5.0              # 并行测试

# 代码质量工具
black==23.11.0                   # 代码格式化
isort==5.12.0                    # 导入排序
flake8==6.1.0                    # 代码检查
mypy==1.7.1                      # 类型检查
bandit==1.7.5                    # 安全检查

# 性能分析
psutil==5.9.6                    # 系统资源监控
memory-profiler==0.61.0          # 内存分析
line-profiler==4.1.1             # 行级性能分析

# =============================================================================
# 🔧 配置和环境管理
# =============================================================================

# 配置管理
python-dotenv==1.0.0             # 环境变量管理
pyyaml==6.0.1                    # YAML配置文件
dynaconf==3.2.4                  # 动态配置管理

# 时间处理
python-dateutil==2.8.2           # 日期时间工具
pytz==2023.3                     # 时区处理

# =============================================================================
# 📦 其他工具库
# =============================================================================

# 异步支持
aiofiles==23.2.1                 # 异步文件操作
asyncpg==0.29.0                  # 异步PostgreSQL驱动

# 数据验证和序列化
marshmallow==3.20.1              # 数据序列化
jsonschema==4.20.0               # JSON模式验证

# 任务队列（可选）
celery==5.3.4                    # 分布式任务队列
kombu==5.3.4                     # 消息传递库

# 工具库
click==8.1.7                     # 命令行界面
rich==13.7.0                     # 终端美化输出
tqdm==4.66.1                     # 进度条

# =============================================================================
# 🔒 安全约束
# =============================================================================

# 确保使用安全版本，避免已知漏洞
certifi>=2023.7.22               # SSL证书
urllib3>=2.0.7                   # HTTP库安全版本
setuptools>=68.0.0               # 构建工具安全版本
wheel>=0.41.0                    # 包管理工具

# =============================================================================
# 📋 版本兼容性说明
# =============================================================================

# Python版本要求: >=3.8,<4.0
# 操作系统: Linux, macOS, Windows
# 架构: x86_64, ARM64
# 
# 安全扫描通过日期: 2024-12-29
# 下次更新计划: 2025-01-29 